dapp/back-end/requirements.txt
aiosmtplib==1.1.7
anyio==3.6.1
asgiref==3.5.2
blinker==1.5
click==8.1.3
colorama==0.4.5
dnspython==2.2.1
email-validator==1.3.0
fastapi== 0.85.1
fastapi-mail==1.2.0
h11==0.14.0
idna==3.4
importlib-metadata==5.0.0
Jinja2==3.1.2
MarkupSafe==2.1.1
pydantic==1.10.2
sniffio==1.3.0
starlette==0.20.4
typing-extensions==4.4.0
uvicorn==0.15.0
zipp==3.9.0
sqlmodel
SQLAlchemy
databases
PyJWT
passlib
stripe
python-multipart
alembic
fastapi-sqlalchemy
pydantic-sqlalchemy
asyncpg
psycopg2
astroid==2.15.3
autopep8==2.0.2
click==8.1.3
codespell==2.2.4
dill==0.3.6
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpickle==3.0.1
lazy-object-proxy==1.9.0
mccabe==0.7.0
p2pnetwork==1.2
platformdirs==3.2.0
pony==0.7.16
pycodestyle==2.10.0
pycryptodome==3.17
pylint==2.17.2
tomlkit==0.11.7
Werkzeug==2.1.2
wrapt==1.15.0
